@@ -27,9 +27,9 @@ function deserializeFromBase64 (obj) {
   }
 
   return {
-    from: bs58.encode(new Buffer(obj.from, 'base64')).toString(),
-    seqno: new Buffer(obj.seqno, 'base64'),
-    data: new Buffer(obj.data, 'base64'),
+    from: bs58.encode(Buffer.from(obj.from, 'base64')).toString(),
+    seqno: Buffer.from(obj.seqno, 'base64'),
+    data: Buffer.from(obj.data, 'base64'),
     topicCIDs: obj.topicIDs || obj.topicCIDs
   }
 }
