@@ -14,14 +14,13 @@ module.exports = function (send, hash, callback) {
   // Retrieve the object and its data in parallel, then produce a DAGNode
   // instance using this information.
   parallel([
-    function get (done) {
+    (done) => {
       send({
         path: 'object/get',
         args: multihash
       }, done)
     },
-
-    function data (done) {
+    (done) => {
       // WORKAROUND: request the object's data separately, since raw bits in JSON
       // are interpreted as UTF-8 and corrupt the data.
       // See https://github.com/ipfs/go-ipfs/issues/1582 for more details.
@@ -29,25 +28,24 @@ module.exports = function (send, hash, callback) {
         path: 'object/data',
         args: multihash
       }, done)
-    }],
+    }
+  ], (err, res) => {
+    if (err) {
+      return callback(err)
+    }
 
-    function done (err, res) {
-      if (err) {
-        return callback(err)
-      }
+    var object = res[0]
+    var stream = res[1]
 
-      var object = res[0]
-      var stream = res[1]
-
-      if (Buffer.isBuffer(stream)) {
-        DAGNode.create(stream, object.Links, callback)
-      } else {
-        streamToValue(stream, (err, data) => {
-          if (err) {
-            return callback(err)
-          }
-          DAGNode.create(data, object.Links, callback)
-        })
-      }
-    })
+    if (Buffer.isBuffer(stream)) {
+      DAGNode.create(stream, object.Links, callback)
+    } else {
+      streamToValue(stream, (err, data) => {
+        if (err) {
+          return callback(err)
+        }
+        DAGNode.create(data, object.Links, callback)
+      })
+    }
+  })
 }
