@@ -11,12 +11,13 @@ const FactoryClient = require('./ipfs-factory/client')
 const invalidArg = 'this/Is/So/Invalid/'
 const validIp4 = '/ip4/104.236.176.52/tcp/4001/ipfs/QmSoLnSGccFuZQJzRadHn95W2CrSFmZuTdDWP8HXaHca9z'
 
-describe('.bootstrap', () => {
+describe('.bootstrap', function () {
+  this.timeout(100 * 1000)
+
   let ipfs
   let fc
 
-  before(function (done) {
-    this.timeout(20 * 1000) // slow CI
+  before((done) => {
     fc = new FactoryClient()
     fc.spawnNode((err, node) => {
       expect(err).to.not.exist()
@@ -31,7 +32,9 @@ describe('.bootstrap', () => {
 
   let peers
 
-  describe('Callback API', () => {
+  describe('Callback API', function () {
+    this.timeout(100 * 1000)
+
     describe('.add', () => {
       it('returns an error when called with an invalid arg', (done) => {
         ipfs.bootstrap.add(invalidArg, (err) => {
@@ -112,7 +115,9 @@ describe('.bootstrap', () => {
     })
   })
 
-  describe('Promise API', () => {
+  describe('Promise API', function () {
+    this.timeout(100 * 1000)
+
     describe('.add', () => {
       it('returns an error when called without args or options', () => {
         return ipfs.bootstrap.add(null)
