@@ -1,20 +1,22 @@
 'use strict'
 
-const gulp = require('gulp')
 const factoryServer = require('./server')
 
 let factory
 
-gulp.task('factory:start', (done) => {
-  factoryServer((err, http) => {
-    if (err) {
-      throw err
-    }
-    factory = http
-    done()
-  })
-})
-
-gulp.task('factory:stop', (done) => {
-  factory.stop(done)
-})
+module.exports = {
+  start (done) {
+    factoryServer((err, http) => {
+      if (err) {
+        return done(err)
+      }
+      factory = http
+      done()
+    })
+  },
+  stop (done) {
+    factory.stop({
+      timeout: 1
+    }, done)
+  }
+}
