@@ -8,12 +8,13 @@ const expect = chai.expect
 chai.use(dirtyChai)
 const FactoryClient = require('./ipfs-factory/client')
 
-describe('.log', () => {
+describe('.log', function () {
+  this.timeout(100 * 1000)
+
   let ipfs
   let fc
 
-  before(function (done) {
-    this.timeout(20 * 1000) // slow CI
+  before((done) => {
     fc = new FactoryClient()
     fc.spawnNode((err, node) => {
       expect(err).to.not.exist()
@@ -24,7 +25,9 @@ describe('.log', () => {
 
   after((done) => fc.dismantle(done))
 
-  describe('Callback API', () => {
+  describe('Callback API', function () {
+    this.timeout(100 * 1000)
+
     it('.log.tail', (done) => {
       const req = ipfs.log.tail((err, res) => {
         expect(err).to.not.exist()
@@ -62,7 +65,9 @@ describe('.log', () => {
     })
   })
 
-  describe('Promise API', () => {
+  describe('Promise API', function () {
+    this.timeout(100 * 1000)
+
     it('.log.tail', () => {
       return ipfs.log.tail()
         .then((res) => {
