@@ -10,15 +10,16 @@ const waterfall = require('async/waterfall')
 const path = require('path')
 const FactoryClient = require('./ipfs-factory/client')
 
-describe('.refs', () => {
+describe('.refs', function () {
+  this.timeout(80 * 1000)
+
   if (!isNode) { return }
 
   let ipfs
   let fc
   let folder
 
-  before(function (done) {
-    this.timeout(20 * 1000) // slow CI
+  before((done) => {
     fc = new FactoryClient()
     waterfall([
       (cb) => fc.spawnNode(cb),
